LayoutTests/platform/ios-simulator/ios/scrolling/overflow-div-scrolling-expected.txt

CONSOLE MESSAGE: line 24: TypeError: undefined is not a function (evaluating 'eventSender.clearTouchPoints()')
Tests that touch events cause an overflow:scroll div to scroll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS targetDiv.scrollLeft is 0
PASS targetDiv.scrollTop is 0
